# [KERNELS] Improve block sizes for batched matmul_ogs with small m/n/k.