(Previously, block sizes could be much bigger than m/n/k.)

Example perf difference:
```
H100:
    B=500000 M=8 N=8 K=8
        >> torch.float16     0.850 ms -> 0.388 ms
        >> torch.bfloat16    0.828 ms -> 0.354 ms
        >> torch.float8_e5m2 0.829 ms -> 0.373 ms
    B=500000 M=16 N=16 K=16
        >> torch.float16     0.791 ms -> 0.381 ms
        >> torch.bfloat16    0.790 ms -> 0.382 ms
        >> torch.float8_e5m2 0.779 ms -> 0.366 ms

GB200:
    B=500000 M=8 N=8 K=8
        >> torch.float16     0.676 ms -> 0.314 ms
        >> torch.bfloat16    0.652 ms -> 0.297 ms
        >> torch.float8_e5m2 0.659 ms -> 0.294 ms
    B=500000 M=16 N=16 K=16
        >> torch.float16     0.622 ms -> 0.305 ms
        >> torch.bfloat16    0.606 ms -> 0.306 ms
        >> torch.float8_e5m2 0.616 ms -> 0.296 ms
```